test: ut
refactor: more

refactor: more

test: ut

refactor: telemetry update

fix: pr comment
