+++ packages/fx-core/resource/package.nls.json
@@ -301,6 +301,10 @@
   "core.createProjectQuestion.projectType.customCopilot.label": "Custom Copilot",
   "core.createProjectQuestion.projectType.customCopilot.title": "App Features Using Teams AI Library",
   "core.createProjectQuestion.projectType.customCopilot.placeholder": "Select an option",
+  "core.createProjectQuestion.projectType.copilotHelp.label": "Don't know how to start? Use Github Copilot Chat",
+  "core.createProjectQuestion.projectType.copilotHelp.detail": "Chat with Github Copilot to get step by step instructions for Teams app development",
+  "core.createProjectQuestion.projectType.copilotGroup.title": "Copilot",
+  "core.createProjectQuestion.projectType.createGroup.title": "Create",
   "core.createProjectQuestion.title": "New Project",
   "core.createProjectQuestion.capability.botMessageExtension.label": "Start with a Bot",
   "core.createProjectQuestion.capability.botMessageExtension.detail": "Create a message extension using Bot Framework",
