+++ packages/vscode-extension/package.nls.json
@@ -134,6 +134,8 @@
     "teamstoolkit.commandsTreeViewProvider.previewDescription": "Debug and preview your Teams app",
     "teamstoolkit.commandsTreeViewProvider.previewTitle": "Preview Your Teams App (F5)",
     "_teamstoolkit.commandsTreeViewProvider.previewTitle.comment": "'F5' is a shortcut key, no need to translate it.",
+    "teamstoolkit.commandsTreeViewProvider.getCopilotHelpTitle": "Get help from Github Copilot",
+    "teamstoolkit.commandsTreeViewProvider.getCopilotHelpDescription": "Chat with Github Copilot to find out what you can do for your current status.",
     "teamstoolkit.commandsTreeViewProvider.provision.blockTooltip": "Unable to run command during provisioning. Try again after provisioning completes.",
     "teamstoolkit.commandsTreeViewProvider.provision.running": "Provisioning in progress...",
     "teamstoolkit.commandsTreeViewProvider.provisionDescription": "Run the 'provision' lifecycle stage in teamsapp.yml file",
@@ -143,8 +145,6 @@
     "teamstoolkit.commandsTreeViewProvider.publishDescription": "Run the 'publish' lifecycle stage in teamsapp.yml file.",
     "teamstoolkit.commandsTreeViewProvider.publishInDevPortalTitle": "Open Developer Portal to Publish",
     "teamstoolkit.commandsTreeViewProvider.publishInDevPortalDescription": "Publish to your org in Developer Portal",
-    "teamstoolkit.commandsTreeViewProvider.invokeChat": "Chat with Teams Agent",
-    "teamstoolkit.commandsTreeViewProvider.invokeChatDescription": "Start chatting with Teams Agent for development",
     "teamstoolkit.commandsTreeViewProvider.publishTitle": "Publish",
     "teamstoolkit.commandsTreeViewProvider.getStartedTitle": "Get Started",
     "teamstoolkit.commandsTreeViewProvider.reportIssuesDescription": "Report any issues and let us know your feedback",
@@ -258,6 +258,13 @@
     "teamstoolkit.localDebug.failedCheckers": "Unable to check: [%s].",
     "teamstoolkit.handlers.askInstallOfficeAddinDependency": "Install dependencies for Office Add-in?",
     "teamstoolkit.handlers.installOfficeAddinDependencyCancelled": "Dependency installation is canceled, but you can install dependencies manually by clicking the 'Development - Check and Install Dependencies' button on the left side.",
+    "teamstoolkit.handlers.askInstallCopilot": "To start with Github Copilot, we recommend you to install the extenstion first.",
+    "teamstoolkit.handlers.askInstallCopilot.install": "Install",
+    "teamstoolkit.handlers.askInstallCopilot.cancel": "Cancel",
+    "teamstoolkit.handlers.installExtension.output": "You need to install %s first.",
+    "teamstoolkit.handlers.installCopilotError": "Cannot install Github Copilot Chat. Please install it from extension marketplace and try again.",
+    "teamstoolkit.handlers.chatTeamsAgentError": "Cannot focus Github Copilot Chat automatically. Please open Github Copilot Chat and start with \"@teams\"",
+    "teamstoolkit.handlers.verifyCopilotExtensionError": "Cannot install Github Copilot Chat. Please install it manually and try again.",
     "teamstoolkit.localDebug.learnMore": "Learn More",
     "teamstoolkit.localDebug.m365TenantHintMessage": "After enrolling your developer tenant in Office 365 Target Release, enrollment may come into effect in couple of days. Click 'Learn More' button for details on setting up dev environment to extend Teams apps across Microsoft 365.",
     "teamstoolkit.localDebug.npmInstallFailedHintMessage": "Task '%s' did not complete successfully. For detailed error information, check '%s' terminal window and to report the issue, click 'Report Issue' button.",
