+++ packages/vscode-extension/src/controls/sampleGallery/SampleGallery.scss
@@ -10,11 +10,6 @@
     margin-left: 24px;
   }
 
-  #head {
-    display: flex;
-    justify-content: space-between;
-  }
-
   #featured-sample-title {
     width: 100%;
     height: 43px;
